% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-tooltip.R
\name{add_field}
\alias{add_field}
\title{Add a field to a Vega tooltip}
\usage{
add_field(tooltip, ...)
}
\arguments{
\item{tooltip}{\code{vega_tooltip} object, created using \code{\link[=vega_tooltip]{vega_tooltip()}}}

\item{...}{other arguments (not supported)}
}
\value{
\code{vega_tooltip} object
}
\description{
TODO: get documentation sorted out - show something working for now
}
\examples{
  plot_basic <-
    alt$Chart(
      r_to_py(mtcars)
    )$encode(
      x = "mpg:Q",
      y = "hp:Q",
      color = "cyl:N"
    )$mark_point()

  tooltip_custom <-
    vega_tooltip() \%>\%
    add_field(field = "mpg", title = "MPG") \%>\%
    add_field(field = "hp", title = "HP")

\dontrun{
  vegalite(plot_basic, tooltip = tooltip_custom)
}
}
