% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{alt}
\alias{alt}
\title{Altair object}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
alt
}
\description{
Uses the reticulate framework to access the Altair API.
}
\details{
The Altair Python library is exposed through the \code{alt} object.
You can create and add to chart using its methods and classes,
as outlined in the
\href{https://altair-viz.github.io/index.html}{Altair Python documentation}.

In this package, use the \code{$} operator wherever you see the \code{.} operator
in Python.
}
\examples{
  plot_basic <-
    alt$Chart(
      r_to_py(mtcars)
    )$encode(
      x = "mpg:Q",
      y = "hp:Q",
      color = "cyl:N"
    )$mark_point()

}
\keyword{datasets}
