% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altair-package.R
\docType{package}
\name{altair-package}
\alias{altair}
\alias{altair-package}
\title{altair: Create and embed Vega-Lite charts using the Altair Python package}
\description{
The goal of altair is to help you build
\href{https://vega.github.io/vega-lite/}{Vega-Lite} visualizations. Using the
\href{https://rstudio.github.io/reticulate/}{reticulate} package, it provides
an interface to the \href{https://altair-viz.github.io}{Altair} Python package.
}
\details{
In this documentation, the capitalized word \strong{Altair} shall refer to the
Python package; the lower-case word \strong{altair} shall refer to this R package.
}
\seealso{
\href{https://vegawidget.github.io/altair/}{altair pkgdown website},
\href{https://altair-viz.github.io/index.html}{Altair Python package},
\href{https://vega.github.io/vega-lite/}{Vega-Lite}
}
\author{
\strong{Maintainer}: Ian Lyttle \email{ijlyttle@me.com} (\href{https://orcid.org/0000-0001-9962-4849}{ORCID})

Authors:
\itemize{
  \item Haley Jeppson \email{hjeppson@iastate.edu}
  \item Altair Developers
}

Other contributors:
\itemize{
  \item Alicia Schep \email{aschep@gmail.com} (\href{https://orcid.org/orcid.org/0000-0002-3915-0618}{ORCID}) [contributor]
  \item Jake Vanderplas (Altair library) [contributor]
  \item Brian Granger (Altair library) [contributor]
}

}
\keyword{internal}
