% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{altair}
\alias{altair}
\title{Altair object}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
altair
}
\description{
Uses the reticulate framework to access the Altair API
}
\keyword{datasets}
