% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{altair_concatenation}
\alias{altair_concatenation}
\alias{|.altair.vegalite.v2.api.TopLevelMixin}
\alias{+.altair.vegalite.v2.api.TopLevelMixin}
\alias{&.altair.vegalite.v2.api.TopLevelMixin}
\title{Altair plot concatenation}
\usage{
\method{|}{altair.vegalite.v2.api.TopLevelMixin}(e1, e2)

\method{+}{altair.vegalite.v2.api.TopLevelMixin}(e1, e2)

\method{&}{altair.vegalite.v2.api.TopLevelMixin}(e1, e2)
}
\arguments{
\item{e1}{An Altair plot object}

\item{e2}{An Altair plot object}
}
\value{
Compound Altair plot object
}
\description{
Altair plots can be concatenated using the following operators:
\code{+}, \code{|}, and \code{&}
}
\examples{

# Layering Charts
base = alt$Chart(r_to_py(beaver1))$encode(
  x = alt$X('time'),
  y = alt$Y('temp', scale = alt$Scale(zero = FALSE))
)

scatter_plot <- base$mark_point()
line_plot <- base$mark_line()

vegalite(scatter_plot + line_plot)
}
