% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{altair_version}
\alias{altair_version}
\title{Installed versions of Altair, Vega, etc.}
\usage{
altair_version()
}
\value{
named \code{list} of version tags
}
\description{
Returns a named list of version tags for Altair, Vega, Vega-Lite,
and Vega-Embed
}
\examples{
if (interactive()) {
  altair_version()
}

}
