% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{altair_versions}
\alias{altair_versions}
\title{Installed versions of Altair, Vega, etc.}
\usage{
altair_versions()
}
\value{
named \code{list} of version tags
}
\description{
Returns a named list of version tags for Altair, Vega, Vega-Lite,
and Vega-Embed
}
\examples{
  altair_versions()
}
