% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt.R
\name{as_chart}
\alias{as_chart}
\title{Create Altair chart from  vegaspec}
\usage{
as_chart(spec)
}
\arguments{
\item{spec}{vegaspec}
}
\value{
altair object
}
\description{
Create Altair chart from  vegaspec
}
\examples{
  as_chart(vegawidget::spec_mtcars)
}
