% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegawidget.R
\name{as_vegaspec.altair.vegalite.v2.api.TopLevelMixin}
\alias{as_vegaspec.altair.vegalite.v2.api.TopLevelMixin}
\title{Coerce to a Vega/Vega-Lite specification}
\usage{
\method{as_vegaspec}{altair.vegalite.v2.api.TopLevelMixin}(spec, ...)
}
\arguments{
\item{spec}{object to be coerced to Vega/Vega-Lite specification}

\item{...}{other args (attempt to future-proof)}
}
\value{
S3 object of class \code{vegaspec}
}
\description{
Talk about how \code{vegaspec} is a thin wrapper to \code{list}. Implemented as JSON.
}
\seealso{
\code{\link[vegawidget:as_vegaspec]{vegawidget::as_vegaspec()}}
}
