% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget-altair.R, R/utils-vegawidget.R
\name{as_vegaspec.altair.vegalite.v4.api.TopLevelMixin}
\alias{as_vegaspec.altair.vegalite.v4.api.TopLevelMixin}
\alias{as_vegaspec}
\title{Coerce to vegaspec}
\usage{
\method{as_vegaspec}{altair.vegalite.v4.api.TopLevelMixin}(spec, ...)
}
\arguments{
\item{spec}{An object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}

\item{...}{Other arguments (attempt to future-proof)}
}
\description{
See \code{vegawidget::\link[vegawidget]{as_vegaspec}} for details.
}
