% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_altair_version}
\alias{check_altair_version}
\title{Check two verision-strings}
\usage{
check_altair_version(version_installed, version_supported)
}
\arguments{
\item{version_installed}{\code{character} vector, installed version -
can be obtained using \code{alt$__version__}}

\item{version_supported}{\code{character} vector, supported version -
can be obtained using \code{getOption("altair.pyhton.version")}}
}
\value{
invisible \code{NULL}, called for side-effects
}
\description{
Given two version-strings, issue an error, warning, message, or do nothing.
}
\examples{
\dontrun{
  version_supported <- "2.0.1"
  # issues error
  check_altair_version("1.2", version_supported)

  # issues warning
  check_altair_version("2.1", version_supported)

  # issues message
  check_altair_version("2.0.0", version_supported)
  check_altair_version("2.0.0rc1", version_supported)

  # does nothing
  check_altair_version("2.0.1", version_supported)
}
}
\keyword{internal}
