% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectifs.R
\name{eqtp.grade}
\alias{eqtp.grade}
\title{Tableau des EQTP par grade.}
\usage{
eqtp.grade(Bulletins = Bulletins.paie, grade = NULL, regexp = NULL,
  libellés = NULL, agr = FALSE, période = NULL)
}
\arguments{
\item{Bulletins}{Base des bulletins de paye, comportant pour l'ensemble de la période.}

\item{grade}{Grade particulier. Tous les grades en l'absence de spécification.}

\item{regexp}{Vecteur de chaîne de caractères représentant des expressions régulières sur les grades. Tous les grades en l'absence de spécification. La casse est ignorée.}

\item{libellés}{Vecteur de libellés des agrégations de grades par expression régulière. Doit avoir la même dimension que le vecteur de regexp.}

\item{agr}{Booléen (défaut FALSE). Si TRUE, l'expression régulière précédente conduit à agréger les grades décrits par le vecteur d'expressions régulières précédent : une ligne par composante du vecteur.}

\item{période}{Vecteur des années considérées.
\enumerate{
  \item{ les variables charactère suivantes :
     \itemize{
         \item \code{Année}
         \item \code{Matricule}
         \item \code{Statut}
         \item \code{Grade}}}
   \item{ les variables numériques :
       \describe{
         \item{\code{quotité}}{réel entre 0 et 1}}}}.}
}
\value{
Un tableau des effectifs mis en forme avec les grades en ligne et autant de colonnes numériques que d'années de période, plus une colonne de libellés.
}
\description{
Elabore un tableau des équivalents temps plein travaillés par grade et par année.
}
\examples{
eqtp.grade()
}
