% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine.R
\name{examine}
\alias{examine}
\title{Examine a specification}
\usage{
examine(listdata = NULL, mode = "tree", modes = c("code", "form", "text",
  "tree", "view"), ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{listdata}{\code{list} or \code{String} data to view.  Although designed for \code{lists}, \code{listdata} can
be any data source that can be rendered into \code{JSON} with \code{jsonlite}.  Alternately,
\code{listdata} could be a \code{String} of valid \code{JSON}.  This might be helpful
when dealing with an API response.}

\item{mode}{\code{string} for the initial view from \code{modes}.  \code{'tree'} is the default.}

\item{modes}{\code{string} \code{c('code', 'form', 'text', 'tree', 'view')} will be the default, since
these are all the modes currently supported by \code{jsoneditor}.}

\item{...}{\code{list} of other options for \code{jsoneditor}.  This is a temporary way
of trying other options in \code{jsoneditor}.  In the future, this will be eliminated
in favor of specific, more self-documenting and helpful arguments.}

\item{width}{integer in pixels defining the width of the \code{div} container.}

\item{height}{integer in pixels defining the height of the \code{div} container.}

\item{elementId}{character to specify valid \code{CSS} id of the
htmlwidget for special situations in which you want a non-random
identifier.}
}
\value{
called for side-effects
}
\description{
Use this function to interactvely examine a specification, using
\code{\link[listviewer:jsonedit]{listviewer::jsonedit()}}. Although this has been adpated to examing a Vega-Lite
tooltip specification, it could also be adapted to examine an Altair
chart specification.
}
