% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_altair}
\alias{has_altair}
\title{Detect Altair installation}
\usage{
has_altair()
}
\value{
\code{logical} indicating success
}
\description{
Diagnostic function to confirm Altair is working
}
\details{
In the future, this will have more diagnostics and useful feedback.
}
\examples{
\dontrun{
  has_altair()
}
}
