% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_altair}
\alias{install_altair}
\title{Install Altair Python package}
\usage{
install_altair(method = "auto", conda = "auto", pip = TRUE, ...)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{pip}{TRUE to use pip (defaults to TRUE)}

\item{...}{Additional arguments passed to \code{\link[=conda_install]{conda_install()}}
or \code{\link[=virtualenv_install]{virtualenv_install()}}.}
}
\value{
invisible \code{NULL}, called for side-effects
}
\description{
This function wraps \code{\link[reticulate:py_install]{reticulate::py_install()}} to provide a straightforward,
consistent method to install the Altair Python package.
}
\details{
This package uses the \link[reticulate:reticulate]{reticulate} package
to make an interface with the \href{https://altair-viz.github.io/index.html}{Altair}
Python package. To promote consistency in usage of \strong{reticulate} among
different R packages, it is
\href{https://rstudio.github.io/reticulate/articles/package.html#installing-python-dependencies}{recommended}
to use a common Python environment, called \code{"r-reticulate"}.

If you have a Conda installation, you can create this environment using
\code{\link[reticulate:conda_install]{reticulate::conda_install()}}, as described in this
\href{https://rstudio.github.io/reticulate/articles/python_packages.html#conda-installation}{reticulate article},
or in this package's \href{https://ijlyttle.github.io/altair/articles/field-guide-python.html#python-env}{Field-Guide to Python Issues}.

At present, version 2 of the Altair Python package is available as a
release candidtate on PyPi, so it can be installed using \code{pip}.
This function will use \code{pip} to install \code{altair==2.0.0rc2}
into your \code{"r-reticulate"} Python environment.
}
\examples{
\dontrun{
  install_altair()
}
}
\seealso{
\href{https://rstudio.github.io/reticulate/articles/package.html}{Using reticulate in an R Package},
\href{https://rstudio.github.io/reticulate/articles/python_packages.html}{reticulate: Installing Python Packages},
\href{https://ijlyttle.github.io/altair/articles/field-guide-python.html#python-env}{altiar: Field-Guide to Python Issues}
}
