% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{JS}
\alias{JS}
\title{Mark character strings as literal JavaScript code}
\usage{
JS(...)
}
\description{
This function is re-exported from the htmlwidgets package.
See \code{\link[htmlwidgets:JS]{htmlwidgets::JS()}} for more details.
}
\seealso{
\code{\link[htmlwidgets:JS]{htmlwidgets::JS()}}
}
\keyword{internal}
