% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegawidget.R
\name{knit_print.altair.vegalite.v2.api.TopLevelMixin}
\alias{knit_print.altair.vegalite.v2.api.TopLevelMixin}
\title{Knit-print method}
\usage{
knit_print.altair.vegalite.v2.api.TopLevelMixin(x, ..., options = NULL)
}
\arguments{
\item{x}{Altair chart object}

\item{...}{other arguments}

\item{options}{\code{list}, knitr options}
}
\description{
Knit-print method
}
