% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R
\name{knit_print.altair.vegalite.v3.api.Chart}
\alias{knit_print.altair.vegalite.v3.api.Chart}
\alias{knit_print.vegaspec}
\title{Knit-print method}
\usage{
knit_print.altair.vegalite.v3.api.Chart(spec, ..., options = NULL)
}
\arguments{
\item{spec}{An object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}

\item{...}{other arguments}

\item{options}{\code{list}, knitr options}
}
\description{
See \code{vegawidget::\link[vegawidget]{knit_print.vegaspec}} for details,
particularly on additional packages that may have to be installed.
}
