% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-embed.R
\name{only_actions}
\alias{only_actions}
\title{Helper function for vega_embed actions}
\usage{
only_actions(export = FALSE, source = FALSE, editor = FALSE)
}
\arguments{
\item{export}{\code{logical}, include "Export As ..." link}

\item{source}{\code{logical}, include "View Source" link}

\item{editor}{\code{logical}, include "Open in Vega Editor" link}
}
\value{
\code{list}
}
\description{
This function can be useful if you want to specify only one
of the links include in the embedding-options.
}
