% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jupyterlab.R
\name{py_mimebundle}
\alias{py_mimebundle}
\alias{repr_text.altair.vegalite.v2.api.Chart}
\alias{repr_vegalite2.altair.vegalite.v2.api.Chart}
\title{Render in JupyterLab}
\usage{
py_mimebundle(obj)

repr_text.altair.vegalite.v2.api.Chart(chart)

repr_vegalite2.altair.vegalite.v2.api.Chart(chart)
}
\arguments{
\item{obj}{object to be mimed}

\item{chart}{chart object}
}
\value{
called for side-effects
}
\description{
Render in JupyterLab
}
\keyword{internal}
