% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altair.R
\docType{data}
\name{pyr_2010_nontit_fph}
\alias{pyr_2010_nontit_fph}
\title{Pyramide des âges des non titulaires de la fonction publique hospitalière au 31 décembre 2010}
\format{Une base de données contenant 55 lignes, dont un entête, et 5 variables :
\describe{
  \item{age}{âge des agents au 31 décembre 2010}
  \item{Hommes}{nombre d'hommes}
  \item{Femmes}{nombre de femmes}
  \item{année.référence}{référence de l'année (première ligne)}
  \item{versant}{versant de la fonction publique (FPT ou FPH)}
}}
\source{
\url{http://www.insee.fr/fr/themes/document.asp?ref_id=ip1442}
}
\usage{
pyr_2010_nontit_fph
}
\description{
Pyramide des âges des non titulaires de la fonction publique hospitalière au 31 décembre 2010
}
\keyword{datasets}
