% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altair.R
\docType{data}
\name{pyr_2012_fpt}
\alias{pyr_2012_fpt}
\title{Pyramide des âges de la fonction publique territoriale au 31 décembre 2012}
\format{Une base de données contenant 55 lignes, dont un entête, et 5 variables :
\describe{
  \item{age}{âge des agents au 31 décembre 2012}
  \item{Hommes}{nombre d'hommes}
  \item{Femmes}{nombre de femmes}
  \item{année.référence}{référence de l'année (première ligne)}
  \item{versant}{versant de la fonction publique (FPT ou FPH)}
}}
\source{
\url{http://www.fonction-publique.gouv.fr/files/files/statistiques/rapports_annuels/2014/xls/vue_Emploi_Public_20140902.xls}
}
\usage{
pyr_2012_fpt
}
\description{
Pyramide des âges de la fonction publique territoriale au 31 décembre 2012
}
\keyword{datasets}

