% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altair.R
\docType{data}
\name{pyr_2013_fph}
\alias{pyr_2013_fph}
\title{Pyramide des âges de la fonction publique hospitalière au 31 décembre 2013}
\format{Une base de données contenant 55 lignes, dont un entête, et 5 variables :
\describe{
  \item{age}{âge des agents au 31 décembre 2013}
  \item{Hommes}{nombre d'hommes}
  \item{Femmes}{nombre de femmes}
  \item{année.référence}{référence de l'année (première ligne)}
  \item{versant}{versant de la fonction publique (FPT ou FPH)}
}}
\source{
\url{http://www.fonction-publique.gouv.fr/files/files/statistiques/rapports_annuels/2015/xls/vue1_emploi.xlsx}
}
\usage{
pyr_2013_fph
}
\description{
Pyramide des âges de la fonction publique hospitalière au 31 décembre 2013
}
\keyword{datasets}
