% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_to_py}
\alias{r_to_py}
\title{Convert an R object to Python}
\description{
This function is re-exported from the reticulate package.
See \code{\link[reticulate:r_to_py]{reticulate::r_to_py()}} for more details.
}
\seealso{
\code{\link[reticulate:r_to_py]{reticulate::r_to_py()}}
}
\keyword{internal}
