% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{renderVegalite}
\alias{renderVegalite}
\title{Render a shiny output for vegalite}
\usage{
renderVegalite(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates an HTML widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Render a shiny output for vegalite
}
