% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jupyterlab.R
\name{repr_text.altair.vegalite.v4.api.Chart}
\alias{repr_text.altair.vegalite.v4.api.Chart}
\title{MIME method for text}
\usage{
\method{repr_text}{altair.vegalite.v4.api.Chart}(obj)
}
\arguments{
\item{obj}{Altair chart object to be rendered}
}
\value{
character
}
\description{
MIME method for text
}
\keyword{internal}
