% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jupyterlab.R
\name{repr_vegalite2.altair.vegalite.v2.api.Chart}
\alias{repr_vegalite2.altair.vegalite.v2.api.Chart}
\title{MIME method for rendering}
\usage{
repr_vegalite2.altair.vegalite.v2.api.Chart(obj)
}
\arguments{
\item{obj}{chart object to be mimed}
}
\value{
list, reflecting JSON spec
}
\description{
MIME method for rendering
}
\keyword{internal}
