% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jupyterlab.R
\name{repr_vegalite3.altair.vegalite.v3.api.Chart}
\alias{repr_vegalite3.altair.vegalite.v3.api.Chart}
\title{MIME method for chart rendering}
\usage{
repr_vegalite3.altair.vegalite.v3.api.Chart(obj)
}
\arguments{
\item{obj}{Altair chart object to be rendered}
}
\value{
list, reflecting JSON spec
}
\description{
MIME method for chart rendering
}
\keyword{internal}
