% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-jupyterlab.R
\name{repr_vegalite4.altair.vegalite.v5.api.TopLevelMixin}
\alias{repr_vegalite4.altair.vegalite.v5.api.TopLevelMixin}
\title{MIME method for chart rendering}
\usage{
\method{repr_vegalite4}{altair.vegalite.v5.api.TopLevelMixin}(obj, ...)
}
\arguments{
\item{obj}{Altair chart object to be rendered}
}
\value{
list, reflecting JSON spec
}
\description{
MIME method for chart rendering
}
\keyword{internal}
