% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R
\docType{data}
\name{spec_mtcars}
\alias{spec_mtcars}
\title{Example vegaspecs}
\format{An object of class \code{vegaspec_unit} (inherits from \code{vegaspec_vega_lite}, \code{vegaspec}, \code{list}) of length 7.}
\usage{
spec_mtcars
}
\description{
See \code{vegawidget::\link[vegawidget]{spec_mtcars}} for details.
}
\keyword{datasets}
