% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt.R
\name{to_altair}
\alias{to_altair}
\title{Create an Altair object from a vegaspec}
\usage{
to_altair(spec)
}
\arguments{
\item{spec}{vegaspec}
}
\value{
altair object
}
\description{
Create an Altair object from a vegaspec
}
