% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{vega_data}
\alias{vega_data}
\title{Vega datasets}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
vega_data
}
\description{
Uses the reticulate framework to access the \code{vega_datasets} Python
package.
}
\details{
Some data are available locally, others require an internet connection.
}
\examples{
# To list available datasets
vega_data$data$list_datasets()

# When accessing a dataset, substitute any "-" in the name with a "_"
head(vega_data$data$sf_temps())

# Metadata are available for each dataset:
vega_data$data$sf_temps$description
vega_data$data$sf_temps$url

# For local datasets, local path is available
vega_data$data$sf_temps$filepath

}
\seealso{
\href{https://github.com/altair-viz/vega_datasets}{Vega datasets documentation}
}
\keyword{datasets}
