% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{vega_embed}
\alias{vega_embed}
\title{Vega embed options}
\usage{
vega_embed(actions = TRUE)
}
\arguments{
\item{actions}{\code{logical} or \code{list} of logicals
Vega-Embed default is all TRUE.}
}
\value{
\code{list}
}
\description{
Vega embed options
}
