% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegalite-htmlwidget.R
\name{vegalite}
\alias{vegalite}
\title{Create a Vega-Lite htmlwidget}
\usage{
vegalite(chart, tooltip = vega_tooltip(), embed = vega_embed(),
  width = NULL, height = NULL)
}
\arguments{
\item{chart}{an Altair plot object}

\item{tooltip}{\code{vega_tooltip} object to specify tooltip -
the default is an empty call to \code{\link[=vega_tooltip]{vega_tooltip()}},
which will result in no tooltip displaying}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and all actions-links being
included}

\item{width}{\code{integer}, if specified, the total rendered width (in pixels)
of the chart - valid only for single-view charts and layered charts}

\item{height}{\code{integer}, if specified, the total rendered height (in pixels)
of the chart - valid only for single-view charts and layered charts}
}
\description{
Use this function to render a chart as an htmlwidget.
}
\details{
To include a tooltip in your chart-rendering, use one of the
\code{\link[=vega_tooltip]{vega_tooltip()}} functions with the \code{tooltip} argument.
Keep in mind that you can include exactly \strong{one} tooltip specification
and exactly \strong{one} chart specification
in a rendering; the tooltip will be applied to the entire chart,
even if it is a compound chart.

To specify embedding-options, use the \code{\link[=vega_embed]{vega_embed()}} function with the
\code{embed} argument. Its most-important options are:
\itemize{
\item \code{renderer}, to specify \code{"canvas"} (default) or \code{"svg"}
\item \code{actions},  to specific action-links
for \code{export}, \code{source}, and \code{editor}
}

If actions is a single \code{TRUE} (default) or \code{FALSE}, all the links
are shown (or not). Use a named list to be more specific, see
\code{\link[=vega_embed]{vega_embed()}} and \code{\link[=only_actions]{only_actions()}}.

This function is called \code{vegalite()} is because it returns an htmlwidget
that uses the Vega-Lite JavaScript library, rather than the
Altair Python package.
}
\examples{
  plot_basic <-
    alt$Chart(
      r_to_py(mtcars)
    )$encode(
      x = "mpg:Q",
      y = "hp:Q",
      color = "cyl:N"
    )$mark_point()

\dontrun{
  vegalite(plot_basic)
}
}
\seealso{
\link{alt}, \code{\link[=vega_tooltip]{vega_tooltip()}}, \code{\link[=vega_embed]{vega_embed()}}, \code{\link[=only_actions]{only_actions()}}
}
