% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{vegaliteOutput}
\alias{vegaliteOutput}
\title{Shiny output for vegalite}
\usage{
vegaliteOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}

\item{height}{Must be a valid CSS unit (like \code{"100\%"},
\code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
string and have \code{"px"} appended.}
}
\description{
Shiny output for vegalite
}
