% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R
\name{vw-add-listeners}
\alias{vw-add-listeners}
\alias{vw_add_event_listener}
\alias{vw_add_signal_listener}
\title{JavaScript listeners for vegawidget}
\description{
See \code{vegawidget::\link[vegawidget]{add-listeners}} for details.
}
