% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R
\name{vw-handlers}
\alias{vw-handlers}
\alias{vw_handler_signal}
\alias{vw_handler_event}
\alias{vw_handler_add_effect}
\title{JavaScript handlers for vegawidget}
\description{
See \code{vegawidget::\link[vegawidget]{vw_handler_signal}},
\code{vegawidget::\link[vegawidget]{vw_handler_add_effect}}, for details.
}
