% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R
\name{vw-shiny-getters}
\alias{vw-shiny-getters}
\alias{vw_shiny_get_event}
\alias{vw_shiny_get_signal}
\title{Shiny getters for vegawidget}
\description{
See \code{vegawidget::\link[vegawidget]{shiny-getters}} for details.
}
