% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R
\name{vw-shiny-setters}
\alias{vw-shiny-setters}
\alias{vw_shiny_set_data}
\alias{vw_shiny_set_signal}
\title{Shiny setters for vegawidget}
\description{
See \code{vegawidget::\link[vegawidget]{shiny-setters}} for details.
}
