% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{vw_as_json}
\alias{vw_as_json}
\title{Coerce vegaspec to JSON}
\usage{
vw_as_json(spec, pretty = TRUE)
}
\arguments{
\item{spec}{An object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}

\item{pretty}{\code{logical} indicates to use pretty (vs. minified) formatting}
}
\value{
\code{jsonlite::json} object
}
\description{
Deprecated, please use \code{vegawidget::\link[vegawidget]{vw_as_json}}.
}
