% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-tooltip.R
\name{add_field}
\alias{add_field}
\title{Add a field to a Vega tooltip}
\usage{
add_field(tooltip, ...)
}
\arguments{
\item{tooltip}{\code{vega_tooltip} object, created using \code{\link[=vega_tooltip]{vega_tooltip()}}}

\item{...}{other arguments (not supported)}
}
\value{
\code{vega_tooltip} object
}
\description{
TODO: get documentation sorted out - show something working for now
}
