% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altair-package.R
\docType{package}
\name{altair}
\alias{altair}
\alias{altair-package}
\title{altair: Create Vega-Lite charts using the Altair Python package}
\description{
altair: Create Vega-Lite charts using the Altair Python package
}
