% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt.R
\name{as_chart}
\alias{as_chart}
\title{Create Altair chart from vegaspec}
\usage{
as_chart(spec)
}
\arguments{
\item{spec}{An object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}
}
\value{
altair object
}
\description{
Create Altair chart from vegaspec
}
\examples{
  if (interactive()) {
    as_chart(vegawidget::spec_mtcars)
  }
}
