% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R
\name{as_vegaspec.altair.vegalite.v2.api.TopLevelMixin}
\alias{as_vegaspec.altair.vegalite.v2.api.TopLevelMixin}
\alias{as_vegaspec}
\title{Coerce to vegaspec}
\usage{
\method{as_vegaspec}{altair.vegalite.v2.api.TopLevelMixin}(spec, ...)
}
\arguments{
\item{spec}{object to be coerced to \code{vegaspec}, a Vega/Vega-Lite specification}

\item{...}{other args (attempt to future-proof)}
}
\value{
S3 object of class \code{vegaspec}
}
\description{
See \code{vegawidget::\link[vegawidget]{as_vegaspec}} for details.
}
