% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_altair}
\alias{install_altair}
\title{Install Altair Python package}
\usage{
install_altair(method = "auto", conda = "auto", pip = TRUE, ...)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
PATH and other conventional install locations).}

\item{pip}{TRUE to use pip (defaults to TRUE)}

\item{...}{Additional arguments passed to \code{\link[=conda_install]{conda_install()}}
or \code{\link[=virtualenv_install]{virtualenv_install()}}.}
}
\value{
invisible \code{NULL}, called for side-effects
}
\description{
Install Altair Python package
}
\examples{
\dontrun{
  install_altair()
}
}
