% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt.R
\name{knit_print.altair.vegalite.v2.api.TopLevelMixin}
\alias{knit_print.altair.vegalite.v2.api.TopLevelMixin}
\title{Knit-print method}
\usage{
knit_print.altair.vegalite.v2.api.TopLevelMixin(x, ..., options = NULL)
}
\arguments{
\item{x}{Altair chart object}

\item{...}{other arguments}

\item{options}{\code{list}, knitr options}
}
\description{
The only effective options are \code{vega.width} and \code{vega.height};
these are passed to \code{\link[=vegawidget]{vegawidget()}} as \code{width} and \code{height},
respectively.
}
\details{
\code{out.width} and \code{out.height} are
coerced to numeric, so it is ineffective to specify a percentage.
}
