% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-embed.R
\name{vega_embed}
\alias{vega_embed}
\title{Vega embed options}
\usage{
vega_embed(renderer = c("canvas", "svg"), actions = TRUE, mode = NULL,
  logLevel = NULL, loader = NULL, onBeforeParse = NULL, width = NULL,
  height = NULL, padding = NULL, config = NULL, editorUrl = NULL,
  sourceHeader = NULL, sourceFooter = NULL, runAsync = NULL)
}
\arguments{
\item{renderer}{\code{character} the renderer to use for the view.
One of \code{"canvas"} (default) or \code{"svg"}.
See \href{https://vega.github.io/vega/docs/api/view/#view_renderer}{Vega docs}
for details.}

\item{actions}{\code{logical} or named vector of logicals, determines if action links
("Export as PNG/SVG", "View Source", "Open in Vega Editor")
are included with the embedded view.
If the value is \code{TRUE} (default), all action links will be shown
and none if the value is \code{FALSE}. This property can be a named vector of
logicals that maps
keys (\code{export}, \code{source}, \code{editor}) to logical values for determining
if each action link should be shown. Unspecified keys will be true by
default. For example, if \code{actions} is
\code{c(export = FALSE, source = TRUE)}, the embedded visualization will
have two links – "View Source" and "Open in Vega Editor".}

\item{mode}{\code{character} if specified, tells Vega-Embed to parse the spec
as vega or vega-lite. Vega-Embed will parse the \code{$schema} url if the mode
is not specified. Vega-Embed will default to vega if neither mode,
nor $schema are specified.}

\item{logLevel}{sets the current log level.
See \href{https://vega.github.io/vega/docs/api/view/#view_logLevel}{Vega docs}
for details.}

\item{loader}{sets a custom Vega loader.
See \href{https://github.com/vega/vega-loader#loader}{Vega docs}
for details.}

\item{onBeforeParse}{\code{JS} function, modifies the spec before being parsed.}

\item{width}{\code{integer} sets the view width in pixels.
See \href{https://vega.github.io/vega/docs/api/view/#view_width}{Vega docs}
for details.
Note that Vega-Lite overrides this option.}

\item{height}{\code{integer}, sets the view height in pixels.
See \href{https://vega.github.io/vega/docs/api/view/#view_height}{Vega docs}
for details.
Note that Vega-Lite overrides this option.}

\item{padding}{\code{list}, sets the view padding in pixels.
See \href{https://vega.github.io/vega/docs/api/view/#view_padding}{Vega docs}
for details.}

\item{config}{\code{character} or \code{list} a URL string** from which to load
a Vega/Vega-Lite or Vega-Lite configuration file, or a \code{list} of
Vega/Vega-Lite configurations to override the default configuration
options. If \code{config} is a URL, it will be subject to standard browser
security restrictions. Typically this URL will point to a file on the same
host and port number as the web page itself.}

\item{editorUrl}{\code{character}, URL at which to open embedded Vega specs
in a Vega editor. Defaults to \code{"http://vega.github.io/editor/"}.
Internally, Vega-Embed uses
\href{https://developer.mozilla.org/en-US/docs/Web/API/Window/postMessage}{HTML5 postMessage}
to pass the specification information to the editor.}

\item{sourceHeader}{\code{character}, HTML to inject into the `` tag of the
page generated by the "View Source" action link. For example, this
can be used to add code for \href{https://highlightjs.org/}{syntax highlighting}.}

\item{sourceFooter}{\code{character}, HTML to inject into the end of the
page generated by the "View Source" action link. The text will be added
immediately before the closing `` tag.}

\item{runAsync}{\code{logical}, use
\href{https://vega.github.io/vega/docs/api/view/#view_runAsync}{runAsync}
instead of \href{https://vega.github.io/vega/docs/api/view/#view_run}{run}.}
}
\value{
\code{list}
}
\description{
Vega embed options
}
