% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega-tooltip.R
\name{vega_tooltip}
\alias{vega_tooltip}
\title{Create a Vega tooltip spec}
\usage{
vega_tooltip(showAllFields = FALSE, fields = data.frame(), delay = 100,
  colorTheme = c("light", "dark"), sort = c("title", "value"))
}
\arguments{
\item{showAllFields}{\code{logical}, indicates which fields to include in tooltip.
\itemize{
\item \code{TRUE}: include all fields in the data.
\item \code{FALSE}: include only those fields specified in \code{fields}.
Use this with empty \code{fields} to suppress tooltip.
\item \code{NULL}: include only those fields encoded in the chart.
}}

\item{fields}{\code{data.frame}, used if \code{showAllFields == FALSE}.
One row for each field to include in tooltip, with variables:
\itemize{
\item \code{field} (\code{character}) The unique name of the field. With Vega-Lite,
this is the field you provided to each encoding channel.
\item \code{title} (\code{character}) A custom title for the field.
\item \code{formatType} (\code{character}) Tells what kind of field this is
(for formatting the field value in the tooltip)
Supported values: \code{"number"}, \code{"time"}, and \code{"string"}.
\item \code{format} (\code{character}) A string specifier for formatting the field value
in the tooltip. If \code{formatType} is \code{"number"}, you can provide a
\href{https://github.com/d3/d3-format#locale_format}{number format string-specifier}.
If \code{formatType} is \code{"time"}, you can provide a
\href{https://github.com/d3/d3-time-format#locale_format}{time format string-specifier}.
If \code{formatType} is \code{"string"}, there is no need to provide a format.
\item \code{aggregate} (\code{character}) If your Vega-Lite visualization has multiple
aggregations of the same field, you can specify the aggregation
to identify the particular aggregated field.
Supported values: \href{https://vega.github.io/vega-lite/docs/aggregate.html#supported-aggregation-operations}{Vega-Lite aggregate operations}
}}

\item{delay}{\code{numeric} number of milliseconds tooltip display
should be delayed.}

\item{colorTheme}{\code{character} color theme picker}

\item{sort}{\code{character} specify sorting fields on either
\code{"title"}, or \code{"value"}, using sorting orders:
\itemize{
\item dates and strings are sorted ascending, numbers descending
\item dates appear first, then numbers, then strings
}}
}
\value{
S3 object with class \code{vega_tooltip}
}
\description{
Leaving out the specification of JavaScript functions for now, because I don't
know how to handle them (yet).
}
\seealso{
\href{https://github.com/vega/vega-tooltip/blob/master/docs/customizing_your_tooltip.md#options}{Vega tooltip documentation}
}
