% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegawidget.R
\name{vegalite}
\alias{vegalite}
\title{Rendering function - deprecated}
\usage{
vegalite(chart, embed = NULL, width = NULL, height = NULL, ...)
}
\arguments{
\item{chart}{an Altair chart object}

\item{embed}{\code{vega_embed} object to specify embedding options -
the default is an empty call to \code{\link[=vega_embed]{vega_embed()}},
which will result in a canvas-rendering and action-links included for
exporting, viewing the Vega-Lite source, and opening the Vega editor.}

\item{width}{\code{integer}, if specified, the total rendered width (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the width in the chart specification}

\item{height}{\code{integer}, if specified, the total rendered height (in pixels)
of the chart - valid only for single-view charts and layered charts;
the default is to use the height in the chart specification}

\item{...}{other arguments}
}
\description{
This function has been renamed to \code{vegawidget()}
}
\keyword{internal}
