% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidgets.R
\name{vegalite}
\alias{vegalite}
\title{Create a vegalite htmlwidget}
\usage{
vegalite(chart, width = NULL, height = NULL)
}
\arguments{
\item{chart}{chart object created using \code{altair$Chart()}}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}
}
\description{
Create a vegalite htmlwidget
}
